requirements.txt
asttokens==2.4.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
customtkinter==5.2.1
darkdetect==0.8.0
exceptiongroup==1.1.3
executing==2.0.0
h11==0.14.0
icecream==2.1.3
idna==3.4
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
Pygments==2.16.1
PySocks==1.7.1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
